Extracted account icon into a separate component
