+++ app/components/topbar/WalletAccountIcon.scss
@@ -1,0 +1,4 @@
+.walletIcon {
+    border: 1px solid lightgrey;
+    border-radius: 6px;
+}
