+++ app/components/topbar/WalletTopbarTitle.js
@@ -6,8 +6,7 @@ import styles from './WalletTopbarTitle.scss';
 import { matchRoute } from '../../utils/routing';
 import { ROUTES } from '../../routes-config';
 import Wallet from '../../domain/Wallet';
-import Blockies from 'react-blockies';
-import tinycolor from 'tinycolor2';
+import WalletAccountIcon from "./WalletAccountIcon";
 
 type Props = {
   wallet: ?Wallet,
@@ -16,42 +15,12 @@ type Props = {
   theme: { identiconSaturationFactor: number },
 };
 
-const mkcolor = (primary, secondary, spots) => ({ primary, secondary, spots });
-const COLORS = [
-  mkcolor('#E1F2FF', '#17D1AA', '#A80B32'),
-  mkcolor('#E1F2FF', '#FA5380', '#0833B2'),
-  mkcolor('#E1F2FF', '#f06ef5', '#0804f7'),
-  mkcolor('#E1F2FF', '#ebb687', '#852d62'),
-  mkcolor('#E1F2FF', '#eda9a5', '#43327a'),
-];
-
-const saturation = (color, factor: number = 10) => {
-  if (factor < 0 || factor > 20) {
-    throw Error("Expected factor between 0 and 20 (default 10)")
-  }
-  const diff = factor - 10;
-  let tcol = tinycolor(color);
-  for (let i = 0; i < Math.abs(diff); i++) {
-    tcol = diff < 0 ?
-      tcol.desaturate()
-      : tcol.saturate();
-  }
-  return tcol.toHexString();
-};
-
 function constructPlate(account, saturationFactor): [string, Component] {
-  const { plate: { hash: accountPlateHash, id: accountPlateId } } = account;
-  const colorIdx = Buffer.from(accountPlateHash, 'hex')[0] % COLORS.length;
-  const color = COLORS[colorIdx];
-  return [accountPlateId, (<div className={styles.divIcon}>
-    <Blockies
-      seed={accountPlateHash}
-      size={7}
-      scale={6}
-      bgColor={saturation(color.primary, saturationFactor)}
-      color={saturation(color.secondary, saturationFactor)}
-      spotColor={saturation(color.spots, saturationFactor)}
-      className={styles.walletIcon}
+  const { plate: { hash, id } } = account;
+  return [id, (<div className={styles.divIcon}>
+    <WalletAccountIcon
+      iconSeed={hash}
+      saturationFactor={saturationFactor}
     />
   </div>)]
 }
